feat(sample): 평균 생산시간(분)을 소수점 지원 double 타입으로 변경
- Sample.java: avgProductionMinutes 필드 및 생성자 파라미터 int → double, getAvgProductionMinutes() 반환 타입 int → double
- SampleService.java: register() avgProductionMinutes 파라미터 int → double
- OrderService.java: totalMinutes 계산에서 불필요한 (double) 캐스트 제거
- SampleController.java: 평균 생산시간 파싱 Integer.parseInt() → Double.parseDouble()로 변경하여 소수점 입력 허용
- SampleTest.java, SampleServiceTest.java: assertEquals에 delta 추가하여 double 비교로 수정

+++ src/main/java/org/example/sampleordersystem/controller/SampleController.java
@@ -37,7 +37,7 @@ public class SampleController {
         String stockStr = view.readLine();
 
         try {
-            int avgMin = Integer.parseInt(avgMinStr);
+            double avgMin = Double.parseDouble(avgMinStr);
             double yield = Double.parseDouble(yieldStr);
             int stock = Integer.parseInt(stockStr);
             sampleService.register(id, name, avgMin, yield, stock);
